refactor: extract lib/site from lib/seo
Move site config and OG image path helpers out of lib/seo/site.ts
into lib/site/config.ts and lib/site/routes.ts.
