+++ app/sitemap.ts
@@ -1,8 +1,8 @@
 import type { MetadataRoute } from "next";
 import { getPostFrontmatterBySlug } from "@/lib/blog/frontmatter";
 import { getAllSlugs } from "@/lib/blog/posts";
-import { config } from "@/lib/seo/site";
 import { buildSitemapEntries } from "@/lib/seo/sitemap-entries";
+import { config } from "@/lib/site/config";
 
 export const dynamic = "force-static";
