+++ lib/site/routes.ts
@@ -1,0 +1,8 @@
+export const ogImagePaths = {
+	home: "/opengraph-image",
+	blog: "/blog/opengraph-image",
+} as const;
+
+export function blogPostOgImagePath(slug: string): string {
+	return `/blog/${slug}/opengraph-image`;
+}
